requirements.txt
gspread
oauth2client
pandas
matplotlib
seaborn
openpyxl
reportlab
google-api-python-client